Refactor AbstractBootstrapInitializer
