+++ shardingsphere-proxy/shardingsphere-proxy-bootstrap/src/main/java/org/apache/shardingsphere/proxy/initializer/impl/AbstractBootstrapInitializer.java
@@ -20,15 +20,23 @@ package org.apache.shardingsphere.proxy.initializer.impl;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.shardingsphere.db.protocol.mysql.constant.MySQLServerInfo;
 import org.apache.shardingsphere.db.protocol.postgresql.constant.PostgreSQLServerInfo;
+import org.apache.shardingsphere.infra.config.RuleConfiguration;
+import org.apache.shardingsphere.infra.config.datasource.DataSourceConfiguration;
 import org.apache.shardingsphere.infra.config.datasource.DataSourceParameter;
+import org.apache.shardingsphere.infra.config.persist.ConfigCenter;
 import org.apache.shardingsphere.infra.config.properties.ConfigurationPropertyKey;
 import org.apache.shardingsphere.infra.context.metadata.MetaDataContexts;
 import org.apache.shardingsphere.infra.context.metadata.MetaDataContextsBuilder;
 import org.apache.shardingsphere.infra.metadata.resource.ShardingSphereResource;
+import org.apache.shardingsphere.infra.yaml.config.pojo.YamlRuleConfiguration;
+import org.apache.shardingsphere.infra.yaml.config.swapper.YamlRuleConfigurationSwapperEngine;
 import org.apache.shardingsphere.proxy.backend.communication.jdbc.datasource.factory.JDBCRawBackendDataSourceFactory;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
 import org.apache.shardingsphere.proxy.config.ProxyConfiguration;
 import org.apache.shardingsphere.proxy.config.YamlProxyConfiguration;
+import org.apache.shardingsphere.proxy.config.util.DataSourceParameterConverter;
+import org.apache.shardingsphere.proxy.config.yaml.YamlProxyRuleConfiguration;
+import org.apache.shardingsphere.proxy.config.yaml.YamlProxyServerConfiguration;
 import org.apache.shardingsphere.proxy.database.DatabaseServerInfo;
 import org.apache.shardingsphere.proxy.frontend.ShardingSphereProxy;
 import org.apache.shardingsphere.proxy.initializer.BootstrapInitializer;
@@ -39,11 +47,13 @@ import org.apache.shardingsphere.transaction.context.impl.StandardTransactionCon
 
 import javax.sql.DataSource;
 import java.sql.SQLException;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
+import java.util.Properties;
 import java.util.stream.Collectors;
 
 /**
@@ -143,4 +153,58 @@ public abstract class AbstractBootstrapInitializer implements BootstrapInitializ
     protected abstract TransactionContexts decorateTransactionContexts(TransactionContexts transactionContexts, String xaTransactionMangerType);
     
     protected abstract void initScalingWorker(YamlProxyConfiguration yamlConfig);
+    
+    protected final void persistConfigurations(final ConfigCenter configCenter, final YamlProxyConfiguration yamlConfig, final boolean overwrite) {
+        YamlProxyServerConfiguration serverConfig = yamlConfig.getServerConfiguration();
+        Map<String, YamlProxyRuleConfiguration> ruleConfigs = yamlConfig.getRuleConfigurations();
+        if (!isEmptyLocalConfiguration(serverConfig, ruleConfigs)) {
+            configCenter.persistConfigurations(getDataSourceConfigurationMap(ruleConfigs),
+                    getSchemaRuleConfigurations(ruleConfigs), getGlobalRuleConfigurations(serverConfig.getRules()), serverConfig.getProps(), overwrite);
+        }
+    }
+    
+    private boolean isEmptyLocalConfiguration(final YamlProxyServerConfiguration serverConfig, final Map<String, YamlProxyRuleConfiguration> ruleConfigs) {
+        return ruleConfigs.isEmpty() && serverConfig.getRules().isEmpty() && serverConfig.getProps().isEmpty();
+    }
+    
+    private Map<String, Map<String, DataSourceConfiguration>> getDataSourceConfigurationMap(final Map<String, YamlProxyRuleConfiguration> ruleConfigs) {
+        Map<String, Map<String, DataSourceConfiguration>> result = new LinkedHashMap<>(ruleConfigs.size(), 1);
+        for (Entry<String, YamlProxyRuleConfiguration> entry : ruleConfigs.entrySet()) {
+            result.put(entry.getKey(),
+                    DataSourceParameterConverter.getDataSourceConfigurationMap(DataSourceParameterConverter.getDataSourceParameterMapFromYamlConfiguration(entry.getValue().getDataSources())));
+        }
+        return result;
+    }
+    
+    private Map<String, Collection<RuleConfiguration>> getSchemaRuleConfigurations(final Map<String, YamlProxyRuleConfiguration> yamlRuleConfigs) {
+        YamlRuleConfigurationSwapperEngine swapperEngine = new YamlRuleConfigurationSwapperEngine();
+        return yamlRuleConfigs.entrySet().stream().collect(Collectors.toMap(Entry::getKey,
+            entry -> swapperEngine.swapToRuleConfigurations(entry.getValue().getRules()), (oldValue, currentValue) -> oldValue, LinkedHashMap::new));
+    }
+    
+    private Collection<RuleConfiguration> getGlobalRuleConfigurations(final Collection<YamlRuleConfiguration> globalRuleConfigs) {
+        return new YamlRuleConfigurationSwapperEngine().swapToRuleConfigurations(globalRuleConfigs);
+    }
+    
+    protected final ProxyConfiguration loadProxyConfiguration(final ConfigCenter configCenter) {
+        Collection<String> schemaNames = configCenter.getSchemaMetaDataService().loadAllNames();
+        Map<String, Map<String, DataSourceParameter>> schemaDataSources = loadDataSourceParametersMap(configCenter, schemaNames);
+        Map<String, Collection<RuleConfiguration>> schemaRuleConfigs = loadSchemaRules(configCenter, schemaNames);
+        Properties props = configCenter.getPropsService().load();
+        // TODO load global rules from config center
+        Collection<RuleConfiguration> globalRuleConfigs = configCenter.getGlobalRuleService().load();
+        return new ProxyConfiguration(schemaDataSources, schemaRuleConfigs, globalRuleConfigs, props);
+    }
+    
+    private Map<String, Map<String, DataSourceParameter>> loadDataSourceParametersMap(final ConfigCenter configCenter, final Collection<String> schemaNames) {
+        return schemaNames.stream()
+            .collect(Collectors.toMap(each -> each,
+                each -> DataSourceParameterConverter.getDataSourceParameterMap(configCenter.getDataSourceService().load(each)),
+                (oldValue, currentValue) -> oldValue, LinkedHashMap::new));
+    }
+    
+    private Map<String, Collection<RuleConfiguration>> loadSchemaRules(final ConfigCenter configCenter, final Collection<String> schemaNames) {
+        return schemaNames.stream().collect(
+                Collectors.toMap(each -> each, each -> configCenter.getSchemaRuleService().load(each), (oldValue, currentValue) -> oldValue, LinkedHashMap::new));
+    }
 }
